# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Skill {
  name: string;
  category: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  source: string;
}

export type WeightedSkillPriority = typeof WeightedSkillPriority[keyof typeof WeightedSkillPriority];


export const WeightedSkillPriority = {
  core: 'core',
  important: 'important',
  additional: 'additional',
} as const;

export interface WeightedSkill {
  name: string;
  weight: number;
  priority: WeightedSkillPriority;
}

export interface Role {
  id: string;
  name: string;
  category: string;
  description: string;
  skills: WeightedSkill[];
}

export interface Salary {
  min: number;
  max: number;
  currency: string;
}

export interface Job {
  id: string;
  title: string;
  company: string;
  city: string;
  workMode: string;
  category: string;
  jobType: string;
  experienceLevel: string;
  description: string;
  salary: Salary;
  skills: WeightedSkill[];
  source: string;
  createdAt: string;
  active: boolean;
}

export type JobMatch = Job & {
  matchScore: number;
  skillMatch: number;
  textSimilarity: number;
  matchedSkills: string[];
  missingSkills: string[];
  explanation: string[];
};

export interface AtsBreakdown {
  keywords: number;
  skills: number;
  sections: number;
  projects: number;
  contact: number;
  formatting: number;
}

export interface AtsAnalysis {
  overallScore: number;
  breakdown: AtsBreakdown;
  matchedKeywords: string[];
  missingKeywords: string[];
  sections: string[];
  recommendations: string[];
}

export interface ResumeAnalyzeInput {
  filename: string;
  /**
     * @minLength 20
     * @maxLength 100000
     */
  text: string;
  targetRoleId: string;
}

export interface RecommendationInput {
  /**
     * @minLength 20
     * @maxLength 100000
     */
  text: string;
  targetRoleId: string;
}

export interface ResumeAnalysis {
  filename: string;
  targetRole: Role;
  skills: Skill[];
  readiness: number;
  matchedRoleSkills: string[];
  missingRoleSkills: WeightedSkill[];
  ats: AtsAnalysis;
  recommendations: JobMatch[];
  improvementPlan: string[];
}

export interface Dashboard {
  name: string;
  targetRole: Role;
  skills: Skill[];
  readiness: number;
  ats: AtsAnalysis;
  recommendations: JobMatch[];
  jobsFound: number;
}

export type ReportsCategoriesItem = {
  name: string;
  count: number;
};

export type ReportsGapsItem = {
  name: string;
  count: number;
};

export type ReportsActivityItem = {
  day: string;
  viewed: number;
  saved: number;
  applied: number;
};

export type ReportsDistributionItem = {
  range: string;
  count: number;
};

export interface Reports {
  jobsViewed: number;
  jobsSaved: number;
  jobsApplied: number;
  averageMatch: number;
  atsScore: number;
  readiness: number;
  categories: ReportsCategoriesItem[];
  gaps: ReportsGapsItem[];
  activity: ReportsActivityItem[];
  distribution: ReportsDistributionItem[];
}

export type GetDashboardParams = {
roleId?: string;
};

export type GetJobsParams = {
q?: string;
category?: string;
location?: string;
jobType?: string;
workMode?: string;
sort?: GetJobsSort;
/**
 * @minimum 1
 * @maximum 300
 */
limit?: number;
};

export type GetJobsSort = typeof GetJobsSort[keyof typeof GetJobsSort];


export const GetJobsSort = {
  'best-match': 'best-match',
  'highest-match': 'highest-match',
  newest: 'newest',
} as const;