# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the AI Career Advisor demo application
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: career
    description: Career analysis and recommendations
  - name: jobs
    description: Curated demo job discovery
  - name: activity
    description: Local demo activity and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [career]
      summary: Get the demo career dashboard
      parameters:
        - name: roleId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard profile and recommended opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /roles:
    get:
      operationId: getRoles
      tags: [career]
      summary: List supported target roles
      responses:
        "200":
          description: Available career roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
  /roles/{roleId}:
    get:
      operationId: getRole
      tags: [career]
      summary: Get role requirements
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Target role and weighted skill requirements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Role not found
  /resume/analyze:
    post:
      operationId: analyzeResume
      tags: [career]
      summary: Analyze extracted resume text against a target role
      description: Resume text is processed for this request and is not persisted by the demo API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalyzeInput"
      responses:
        "200":
          description: Skill, ATS, readiness, and matching analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
        "400":
          description: Invalid or empty resume text
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Search and filter the synthetic demo job dataset
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
        - name: workMode
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [best-match, highest-match, newest]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 300
      responses:
        "200":
          description: Matching demo jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a demo job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
  /recommendations:
    post:
      operationId: getRecommendations
      tags: [jobs]
      summary: Rank demo jobs against extracted resume text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "200":
          description: Ranked jobs with transparent match explanations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobMatch"
  /reports:
    get:
      operationId: getReports
      tags: [activity]
      summary: Get demo career analytics
      responses:
        "200":
          description: Career and demo activity analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reports"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Skill:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
      required: [name, category, confidence, source]
    WeightedSkill:
      type: object
      properties:
        name:
          type: string
        weight:
          type: number
        priority:
          type: string
          enum: [core, important, additional]
      required: [name, weight, priority]
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/WeightedSkill"
      required: [id, name, category, description, skills]
    Salary:
      type: object
      properties:
        min:
          type: integer
        max:
          type: integer
        currency:
          type: string
      required: [min, max, currency]
    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        company:
          type: string
        city:
          type: string
        workMode:
          type: string
        category:
          type: string
        jobType:
          type: string
        experienceLevel:
          type: string
        description:
          type: string
        salary:
          $ref: "#/components/schemas/Salary"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/WeightedSkill"
        source:
          type: string
        createdAt:
          type: string
        active:
          type: boolean
      required:
        [id, title, company, city, workMode, category, jobType, experienceLevel, description, salary, skills, source, createdAt, active]
    JobMatch:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            matchScore:
              type: integer
            skillMatch:
              type: integer
            textSimilarity:
              type: integer
            matchedSkills:
              type: array
              items:
                type: string
            missingSkills:
              type: array
              items:
                type: string
            explanation:
              type: array
              items:
                type: string
          required: [matchScore, skillMatch, textSimilarity, matchedSkills, missingSkills, explanation]
    AtsBreakdown:
      type: object
      properties:
        keywords:
          type: integer
        skills:
          type: integer
        sections:
          type: integer
        projects:
          type: integer
        contact:
          type: integer
        formatting:
          type: integer
      required: [keywords, skills, sections, projects, contact, formatting]
    AtsAnalysis:
      type: object
      properties:
        overallScore:
          type: integer
        breakdown:
          $ref: "#/components/schemas/AtsBreakdown"
        matchedKeywords:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
        sections:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
      required: [overallScore, breakdown, matchedKeywords, missingKeywords, sections, recommendations]
    ResumeAnalyzeInput:
      type: object
      properties:
        filename:
          type: string
        text:
          type: string
          minLength: 20
          maxLength: 100000
        targetRoleId:
          type: string
      required: [filename, text, targetRoleId]
    RecommendationInput:
      type: object
      properties:
        text:
          type: string
          minLength: 20
          maxLength: 100000
        targetRoleId:
          type: string
      required: [text, targetRoleId]
    ResumeAnalysis:
      type: object
      properties:
        filename:
          type: string
        targetRole:
          $ref: "#/components/schemas/Role"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        readiness:
          type: integer
        matchedRoleSkills:
          type: array
          items:
            type: string
        missingRoleSkills:
          type: array
          items:
            $ref: "#/components/schemas/WeightedSkill"
        ats:
          $ref: "#/components/schemas/AtsAnalysis"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/JobMatch"
        improvementPlan:
          type: array
          items:
            type: string
      required: [filename, targetRole, skills, readiness, matchedRoleSkills, missingRoleSkills, ats, recommendations, improvementPlan]
    Dashboard:
      type: object
      properties:
        name:
          type: string
        targetRole:
          $ref: "#/components/schemas/Role"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        readiness:
          type: integer
        ats:
          $ref: "#/components/schemas/AtsAnalysis"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/JobMatch"
        jobsFound:
          type: integer
      required: [name, targetRole, skills, readiness, ats, recommendations, jobsFound]
    Reports:
      type: object
      properties:
        jobsViewed:
          type: integer
        jobsSaved:
          type: integer
        jobsApplied:
          type: integer
        averageMatch:
          type: integer
        atsScore:
          type: integer
        readiness:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
            required: [name, count]
        gaps:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
            required: [name, count]
        activity:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              viewed:
                type: integer
              saved:
                type: integer
              applied:
                type: integer
            required: [day, viewed, saved, applied]
        distribution:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
              count:
                type: integer
            required: [range, count]
      required: [jobsViewed, jobsSaved, jobsApplied, averageMatch, atsScore, readiness, categories, gaps, activity, distribution]