# lib/api-zod/src/generated/types/atsAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */
import type { AtsBreakdown } from './atsBreakdown';

export interface AtsAnalysis {
  overallScore: number;
  breakdown: AtsBreakdown;
  matchedKeywords: string[];
  missingKeywords: string[];
  sections: string[];
  recommendations: string[];
}