# lib/api-zod/src/generated/types/atsBreakdown.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */

export interface AtsBreakdown {
  keywords: number;
  skills: number;
  sections: number;
  projects: number;
  contact: number;
  formatting: number;
}