# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}