# lib/api-zod/src/generated/types/recommendationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */

export interface RecommendationInput {
  /**
     * @minLength 20
     * @maxLength 100000
     */
  text: string;
  targetRoleId: string;
}