# lib/api-zod/src/generated/types/reports.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */
import type { ReportsActivityItem } from './reportsActivityItem';
import type { ReportsCategoriesItem } from './reportsCategoriesItem';
import type { ReportsDistributionItem } from './reportsDistributionItem';
import type { ReportsGapsItem } from './reportsGapsItem';

export interface Reports {
  jobsViewed: number;
  jobsSaved: number;
  jobsApplied: number;
  averageMatch: number;
  atsScore: number;
  readiness: number;
  categories: ReportsCategoriesItem[];
  gaps: ReportsGapsItem[];
  activity: ReportsActivityItem[];
  distribution: ReportsDistributionItem[];
}