# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */
import type { WeightedSkill } from './weightedSkill';

export interface Role {
  id: string;
  name: string;
  category: string;
  description: string;
  skills: WeightedSkill[];
}