# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  name: string;
  category: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  source: string;
}