# lib/api-zod/src/generated/types/weightedSkill.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the AI Career Advisor demo application
 * OpenAPI spec version: 0.1.0
 */
import type { WeightedSkillPriority } from './weightedSkillPriority';

export interface WeightedSkill {
  name: string;
  weight: number;
  priority: WeightedSkillPriority;
}